+++ s/demo/components/flat-viewer.ts
@@ -6,7 +6,7 @@ import {DemoView} from "../views/demo_view.js"
 import {common_styles} from "../common-styles.js"
 import {QuickElement} from "../../quick/element.js"
 
-export const FlatViewer = (context: Context) => class extends QuickElement {
+export const FlatViewer = (_: Context) => class extends QuickElement {
 	static styles = common_styles
 
 	render() {